Tidied the code up
