+++ WBooks/Library/BookCell.swift
@@ -9,15 +9,19 @@
 import UIKit
 import WolmoCore
 
-class CustomCell: UITableViewCell, NibLoadable {
+class BookCell: UITableViewCell, NibLoadable {
 
     @IBOutlet weak var title: UILabel!
     @IBOutlet weak var author: UILabel!
     @IBOutlet weak var cover: UIImageView!
+    @IBOutlet weak var viewBackground: UIView!
     
-    func setBook(book: Book){
+    func setBook(book: Book) {
         title.text = book.title
         author.text = book.author
         cover.image = book.cover
+        self.backgroundColor = UIColor.clear
+        viewBackground.layer.cornerRadius = 10
+        viewBackground.clipsToBounds = true
     }
 }
