+++ WBooks/Library/BookCell.xib
@@ -1,0 +1,77 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="14460.31" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina4_7" orientation="portrait">
+        <adaptation id="fullscreen"/>
+    </device>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14460.20"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <view contentMode="scaleToFill" id="mtk-Y5-Qpt" customClass="BookCell" customModule="WBooks" customModuleProvider="target">
+            <rect key="frame" x="0.0" y="0.0" width="375" height="100"/>
+            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+            <subviews>
+                <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="ZFB-PQ-3wH">
+                    <rect key="frame" x="0.0" y="10" width="375" height="90"/>
+                    <subviews>
+                        <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="YNY-ko-4FS">
+                            <rect key="frame" x="124" y="7" width="47.5" height="21.5"/>
+                            <fontDescription key="fontDescription" type="boldSystem" pointSize="18"/>
+                            <nil key="textColor"/>
+                            <nil key="highlightedColor"/>
+                        </label>
+                        <imageView userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="zig-sM-cQ4">
+                            <rect key="frame" x="0.0" y="7.5" width="144" height="75"/>
+                            <constraints>
+                                <constraint firstAttribute="height" constant="75" id="8XJ-jr-jbP"/>
+                                <constraint firstAttribute="width" secondItem="zig-sM-cQ4" secondAttribute="height" multiplier="48:25" id="iYG-MU-F2Y"/>
+                            </constraints>
+                        </imageView>
+                        <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="HCQ-ml-L9Q">
+                            <rect key="frame" x="124" y="64" width="37.5" height="18"/>
+                            <fontDescription key="fontDescription" type="system" pointSize="15"/>
+                            <nil key="textColor"/>
+                            <nil key="highlightedColor"/>
+                        </label>
+                    </subviews>
+                    <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                    <constraints>
+                        <constraint firstItem="zig-sM-cQ4" firstAttribute="centerY" secondItem="ZFB-PQ-3wH" secondAttribute="centerY" id="1Sq-4h-5F6"/>
+                        <constraint firstItem="YNY-ko-4FS" firstAttribute="top" secondItem="lBb-fi-c0a" secondAttribute="top" constant="7" id="BA0-JF-lMn"/>
+                        <constraint firstItem="YNY-ko-4FS" firstAttribute="width" relation="lessThanOrEqual" secondItem="ZFB-PQ-3wH" secondAttribute="width" multiplier="0.6" id="GGM-JG-Npb"/>
+                        <constraint firstItem="HCQ-ml-L9Q" firstAttribute="width" relation="lessThanOrEqual" secondItem="ZFB-PQ-3wH" secondAttribute="width" multiplier="0.6" id="GR5-KV-kTl"/>
+                        <constraint firstItem="HCQ-ml-L9Q" firstAttribute="leading" secondItem="zig-sM-cQ4" secondAttribute="trailing" constant="-20" id="VOL-fH-a8A"/>
+                        <constraint firstItem="zig-sM-cQ4" firstAttribute="leading" secondItem="ZFB-PQ-3wH" secondAttribute="leading" id="W3P-dg-NAy"/>
+                        <constraint firstItem="YNY-ko-4FS" firstAttribute="leading" secondItem="zig-sM-cQ4" secondAttribute="trailing" constant="-20" id="Yoz-cv-OmV"/>
+                        <constraint firstItem="lBb-fi-c0a" firstAttribute="bottom" secondItem="HCQ-ml-L9Q" secondAttribute="bottom" constant="8" id="iEg-cX-Aoq"/>
+                    </constraints>
+                    <viewLayoutGuide key="safeArea" id="lBb-fi-c0a"/>
+                </view>
+            </subviews>
+            <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+            <color key="tintColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+            <constraints>
+                <constraint firstAttribute="bottom" secondItem="ZFB-PQ-3wH" secondAttribute="bottom" id="3mk-Uc-JO3"/>
+                <constraint firstAttribute="trailing" secondItem="ZFB-PQ-3wH" secondAttribute="trailing" id="ISY-6r-fcC"/>
+                <constraint firstItem="ZFB-PQ-3wH" firstAttribute="top" secondItem="mtk-Y5-Qpt" secondAttribute="top" constant="10" id="hHk-9P-yZC"/>
+                <constraint firstItem="ZFB-PQ-3wH" firstAttribute="leading" secondItem="mtk-Y5-Qpt" secondAttribute="leading" id="xha-Pp-iLG"/>
+            </constraints>
+            <nil key="simulatedTopBarMetrics"/>
+            <nil key="simulatedBottomBarMetrics"/>
+            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
+            <viewLayoutGuide key="safeArea" id="P5V-Co-KO0"/>
+            <connections>
+                <outlet property="author" destination="HCQ-ml-L9Q" id="tMZ-0F-tsb"/>
+                <outlet property="cover" destination="zig-sM-cQ4" id="tam-rl-TML"/>
+                <outlet property="title" destination="YNY-ko-4FS" id="c93-QY-bxC"/>
+                <outlet property="viewBackground" destination="ZFB-PQ-3wH" id="LPL-bn-DjR"/>
+            </connections>
+            <point key="canvasLocation" x="-12" y="-177.21139430284859"/>
+        </view>
+    </objects>
+</document>
