+++ WBooks/Library/MainMenuView.xib
@@ -18,12 +18,12 @@
             <subviews>
                 <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" style="plain" separatorStyle="none" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="28" sectionFooterHeight="28" translatesAutoresizingMaskIntoConstraints="NO" id="UlS-Q4-Wsa">
                     <rect key="frame" x="22" y="20" width="331" height="647"/>
-                    <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                    <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                     <color key="separatorColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                     <inset key="separatorInset" minX="15" minY="0.0" maxX="15" maxY="0.0"/>
                 </tableView>
             </subviews>
-            <color key="backgroundColor" red="0.92471915483474731" green="0.96637803316116333" blue="0.97605592012405396" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+            <color key="backgroundColor" red="0.9247191548" green="0.96637803320000004" blue="0.97605592009999997" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
             <constraints>
                 <constraint firstItem="UlS-Q4-Wsa" firstAttribute="leading" secondItem="iN0-l3-epB" secondAttribute="leading" constant="22" id="1GA-Ai-hUo"/>
                 <constraint firstAttribute="trailing" secondItem="UlS-Q4-Wsa" secondAttribute="trailing" constant="22" id="W7d-xz-2Ac"/>
