+++ WBooks/StartingScreen/StartingScreenZeroView.swift
@@ -9,11 +9,8 @@
 import UIKit
 import WolmoCore
 
-
 class StartingScreenZeroView: UIView, NibLoadable {
     
     @IBOutlet weak var signInGoogleButton: UIButton!
     
 }
-
-
